Interactions with Objects
Took out some of the previous models in the 3D models folder inside the
resouces folder.
Added the interaction script to other appliances, using the sink
Interaction script, so that each appliance will spawn a blue ball when
clicked.
Blender Models are stored in the blender models folder, and the actual
prefabs will be used in the  3D models folder